// src/Types/AdministriveStaff.ts
export type AdministriveStaff = {
    userId: string;
    firstName: string;
    lastName: string;
    parentName: string;
    birthday: string;
    address: string;
    city: string;
    role: string;
    age: string;
    email: string;
    gender: string;
    personalNumber: string;
    profilePictureUrl: string;
    telephone: string;
    country: string;
    zipCode: string;
    administriveStaffId: string;
  };
  